+++ fabric/src/main/java/me/adoloveschicken/entombed/EntombedFabric.java
@@ -1,14 +1,23 @@
 package me.adoloveschicken.entombed;
 
+import com.mojang.brigadier.CommandDispatcher;
 import me.adoloveschicken.entombed.block.GravestoneBlockEntity;
 import me.adoloveschicken.entombed.block.ModBlockEntities;
 import me.adoloveschicken.entombed.block.ModBlocks;
+import me.adoloveschicken.entombed.command.TombCommand;
 import me.adoloveschicken.entombed.event.FabricDeathHandler;
 import me.adoloveschicken.entombed.integration.accessory.AccessoriesHandler;
 import me.adoloveschicken.entombed.integration.trinkets.TrinketsHandler;
 import me.adoloveschicken.entombed.item.ModItems;
+import me.adoloveschicken.entombed.storage.GraveIndex;
+import me.adoloveschicken.entombed.storage.GraveStorageManager;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.world.level.storage.LevelResource;
+
+import java.nio.file.Path;
 
 public class EntombedFabric implements ModInitializer {
     @Override
@@ -25,5 +34,16 @@ public class EntombedFabric implements ModInitializer {
         } else if (FabricLoader.getInstance().isModLoaded("accessories")) {
             GravestoneBlockEntity.setGlobalAccessoryHandler(new AccessoriesHandler());
         }
+
+        ServerLifecycleEvents.SERVER_STARTING.register(server ->
+        {
+            Path root = server.getWorldPath(LevelResource.ROOT);
+            GraveStorageManager.initialise(root);
+            GraveIndex.initialise(root);
+        });
+
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> {
+            TombCommand.register(dispatcher);
+        });
     }
 }