+++ neoforge/src/main/java/me/adoloveschicken/entombed/integration/henkelmax/HenkelMaxMigrator.java
@@ -5,8 +5,10 @@ import me.adoloveschicken.entombed.block.GravestoneBlock;
 import me.adoloveschicken.entombed.block.GravestoneBlockEntity;
 import me.adoloveschicken.entombed.block.ModBlocks;
 import me.adoloveschicken.entombed.integration.sable.SableGravePositionHandler;
+import me.adoloveschicken.entombed.storage.GraveStorageManager;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.core.NonNullList;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.*;
@@ -15,6 +17,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.storage.LevelResource;
@@ -34,6 +37,29 @@ public class HenkelMaxMigrator {
     private static final int ARMOR_SLOT_OFFSET = 36;
     private static final int OFFHAND_SLOT = 40;
     private static boolean hasMigrated = false;
+
+    public static CompoundTag migrate(ListTag items, ServerLevel level) {
+        ListTag curiosTagList = new ListTag();
+        for (int i = 0; i < items.size(); i++) {
+            CompoundTag itemEntry = items.getCompound(i);
+            int stackSize = itemEntry.getInt("count");
+            String itemID = itemEntry.getString("id");
+            CompoundTag slotData = itemEntry.getCompound("components").getCompound("baguettelib:curio_slot_data");
+            if (!slotData.isEmpty()) {
+                CompoundTag curiosEntry = new CompoundTag();
+                String slotType = slotData.getString("slotType");
+                int slotIndex = slotData.getInt("slotIndex");
+                ItemStack itemStack = new ItemStack(BuiltInRegistries.ITEM.get(ResourceLocation.parse(itemID)), stackSize);
+                curiosEntry.putString("SlotType", slotType);
+                curiosEntry.putInt("SlotIndex", slotIndex);
+                curiosEntry.put("Item", itemStack.save(level.registryAccess()));
+                curiosTagList.add(curiosEntry);
+            }
+        }
+        CompoundTag extraTag = new CompoundTag();
+        extraTag.put("CurioItems", curiosTagList);
+        return extraTag;
+    }
 
     public static void migrate(MinecraftServer server) {
         Path deathsFolder = server.getWorldPath(LevelResource.ROOT).resolve("deaths");
@@ -83,13 +109,24 @@ public class HenkelMaxMigrator {
         if (level.getBlockEntity(gravePos) instanceof GravestoneBlockEntity grave) {
             grave.setOwnerUUID(ownerUUID);
             grave.setOwnerName(ownerName);
-            loadInventory(deathData.getList("MainInventory", Tag.TAG_COMPOUND), 0, grave, level);
-            loadInventory(deathData.getList("ArmorInventory", Tag.TAG_COMPOUND), ARMOR_SLOT_OFFSET, grave, level);
-            loadInventory(deathData.getList("OffHandInventory", Tag.TAG_COMPOUND), OFFHAND_SLOT, grave, level);
+
+            NonNullList<ItemStack> itemStacks = NonNullList.withSize(GravestoneBlockEntity.INVENTORY_SIZE, ItemStack.EMPTY);
+            loadInventory(deathData.getList("MainInventory", Tag.TAG_COMPOUND), 0, itemStacks);
+            loadInventory(deathData.getList("ArmorInventory", Tag.TAG_COMPOUND), ARMOR_SLOT_OFFSET, itemStacks);
+            loadInventory(deathData.getList("OffHandInventory", Tag.TAG_COMPOUND), OFFHAND_SLOT, itemStacks);
+
+            CompoundTag extraInventoriesTag = new CompoundTag();
             if (ModList.get().isLoaded("curios")) {
-                HenkelMaxCuriosMigrator.migrate(deathData.getList("Items", Tag.TAG_COMPOUND), grave, level);
+                extraInventoriesTag = migrate(deathData.getList("Items", Tag.TAG_COMPOUND), level);
             }
 
+            UUID graveID = UUID.randomUUID();
+            CompoundTag graveData = new CompoundTag();
+            ContainerHelper.saveAllItems(graveData, itemStacks, level.registryAccess());
+            graveData.put("ModExtras", extraInventoriesTag);
+            GraveStorageManager.saveGrave(graveID, graveData);
+
+            grave.setGraveID(graveID);
             grave.setChanged();
             try {
                 Files.move(datFile, datFile.resolveSibling(datFile.getFileName() + ".migrated"));
@@ -99,14 +136,14 @@ public class HenkelMaxMigrator {
         }
     }
 
-    private static void loadInventory(ListTag inventory, int slotOffset, GravestoneBlockEntity grave, ServerLevel level) {
+    private static void loadInventory(ListTag inventory, int slotOffset, NonNullList<ItemStack> itemStacks) {
         for (int i = 0; i < inventory.size(); i++) {
             CompoundTag itemEntry = inventory.getCompound(i);
             int slotIndex = itemEntry.getInt("Slot") + slotOffset;
             int stackSize = itemEntry.getInt("count");
             String itemID = itemEntry.getString("id");
             ItemStack itemStack = new ItemStack(BuiltInRegistries.ITEM.get(ResourceLocation.parse(itemID)), stackSize);
-            grave.setItemInSlot(slotIndex, itemStack);
+            itemStacks.set(slotIndex, itemStack);
         }
     }